# docs/swagger.yaml
basePath: /api
definitions:
...
  models.StyleImage:
    properties:
      createdAt:
        description: 创建时间
        type: string
      description:
        description: 样式图描述
        type: string
      filename:
        description: 文件名
        type: string
      id:
        type: string
      name:
        description: 样式图名称
        type: string
      path:
        description: 文件路径
        type: string
      pointId:
        description: 点位一对一绑定
        type: string
      updatedAt:
        description: 更新时间
        type: string
    type: object
host: localhost:3000
info:
  contact:
    email: support@foreignscan.com
    name: API Support
    url: http://www.foreignscan.com/support
  description: ForeignScan后端API服务
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: ForeignScan API
  version: "1.0"
paths:
...
  /style-images:
    get:
      consumes:
      - application/json
      description: 获取系统中所有可用的样式图片列表
      produces:
      - application/json
      responses:
        "200":
          description: 成功获取样式图片列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取所有样式图片
      tags:
      - style-images
    post:
      consumes:
      - multipart/form-data
      description: 上传或替换点位对照图（point 一对一绑定）
      parameters:
      - description: 点位ID
        in: formData
        name: pointId
        required: true
        type: string
      - description: 样式图文件
        in: formData
        name: file
        required: true
        type: file
      - description: 样式图名称
        in: formData
        name: name
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: 成功上传样式图
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties: true
            type: object
      summary: 上传点位对照图
      tags:
      - style-images
  /style-images/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的样式图及其文件
      parameters:
      - description: 样式图ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: 成功删除样式图
        "404":
          description: 样式图不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties: true
            type: object
      summary: 删除样式图
      tags:
      - style-images
    get:
      consumes:
      - application/json
      description: 根据ID获取特定样式图片的详细信息
      parameters:
      - description: 样式图ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功获取样式图
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 样式图不存在
          schema:
            additionalProperties: true
            type: object
      summary: 获取单个样式图
      tags:
      - style-images
    put:
      consumes:
      - application/json
      description: 更新已存在的样式图信息
      parameters:
      - description: 样式图ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新的样式图信息
        in: body
        name: styleImage
        required: true
        schema:
          $ref: '#/definitions/models.StyleImage'
      produces:
      - application/json
      responses:
        "200":
          description: 成功更新样式图
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 样式图不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties: true
            type: object
      summary: 更新样式图
      tags:
      - style-images
  /style-images/point/{pointId}:
    get:
      consumes:
      - application/json
      description: 每个点位仅允许绑定一张样式图
      parameters:
      - description: 点位ID
        in: path
        name: pointId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功获取样式图片
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 未绑定样式图片
          schema:
            additionalProperties: true
            type: object
      summary: 获取指定点位的样式图
      tags:
      - style-images
...
schemes:
- http
swagger: "2.0"